// src/services/openapi/models/UserListQueryPagedListResult.ts
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */

import type { ResultCode } from './ResultCode';
import type { UserListQueryPagedList } from './UserListQueryPagedList';

export type UserListQueryPagedListResult = {
    readonly isFailure?: boolean;
    readonly isSuccess?: boolean;
    code?: ResultCode;
    readonly errors?: Array<string> | null;
    readonly singleError?: string | null;
    value?: UserListQueryPagedList;
};